requirements.txt
gunicorn
TwitterAPI
elasticsearch>=1.0.0,<2.0.0
pyyaml
